fix error

+++ train3.py
@@ -111,7 +111,7 @@ def eval(model, data_name, save_dir, scale_factor=4, config=None):
         # new_h, new_w = h - h % self.args.size_must_mode, w - w % self.args.size_must_mode
         # gt = gt[:new_h, :new_w, :]
         gt_tensor = utils.numpy2tensor(gt).cuda()
-        gt_tensor, pad = utils.pad_img(gt_tensor, 24*scale_factor)#self.args.size_must_mode*self.args.scale)
+        gt_tensor, pad = utils.pad_img(gt_tensor, int(24*scale_factor))#self.args.size_must_mode*self.args.scale)
         _,_, new_h, new_w = gt_tensor.size()
         # input_tensor = core.imresize(gt_tensor, scale=1/scale_factor)
         # blurred_tensor = core.imresize(input_tensor, scale=scale_factor)
